# openapi/legal-rag-openapi.yaml
openapi: 3.1.0
info:
  title: Legal Agentic RAG Internal Platform API
  version: 0.1.0
  description: Spec-first API for corpus ingestion, QA runtime, evaluation, gold annotation,
    and synthetic dataset generation.
servers:
- url: https://api.example.internal
...
paths:
...
  /v1/runs/{runId}:
    get:
      tags:
      - Runs
      summary: Get one inference run
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    format: uuid
                  dataset_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  question_count:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
                required:
                - run_id
                - dataset_id
                - status
                - question_count
                - created_at
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Inference run UUID
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []